// ditari-app/src/app/models/nxenesi.ts
export interface Nxenesi {
    id?: string;
    emri: string;
    mbiemri: string;
    datelindja: string;
    photo?: string;
    rruga: string;
    qyteti: string;
    numriKontaktues: string;
    email?: string;
    nrLiberAme: string;
    prindi?: number;
    klasaID?: number;
    roli?:number;
}